meditech-system/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
qrcode[pil]==7.4.2
pillow==10.2.0
python-barcode==0.15.1
alembic==1.13.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
